Add AbstractRequestExchange for base exchange tasks
